Revert "Moved the Profile class from profile.py to apio_context.py."
I intended to move project.py, not profile.py. ;-)

This reverts commit bfa6cf45fb8079fca9d3d4872cb92d03e4939d5f.
